trying 1 -

+++ src/evolve.py
@@ -46,8 +46,8 @@ def evolve(config_file):
 
 
     if (rank==0):
+        util.cluster_print(configs['output_directory'], "\nDone. Exiting Evolution.\n")
         comm.Abort()
-        print("\nMaster Thread Exiting Evolution.\n")
 
 
 
@@ -85,9 +85,12 @@ def init_sim(configs, num_sims, sim_num, orig_output_dir, rank):
 def close_out_mult_sims(num_sims, orig_output_dir, configs):
     extract_and_combine(orig_output_dir, num_sims)
     plot_nets.all_plots(configs, indiv_plots=False, orig_output_directory=orig_output_dir)
-    for i in range(num_sims-1):
-        if (os.path.exists(orig_output_dir + "sim_" + str(i))):
-            shutil.rmtree(orig_output_dir + "sim_" + str(i)) #clean up, leave last run as sample
+
+    rm_base_dirs = False
+    if rm_base_dirs:
+        for i in range(num_sims-1):
+            if (os.path.exists(orig_output_dir + "sim_" + str(i))):
+                shutil.rmtree(orig_output_dir + "sim_" + str(i)) #clean up, leave last run as sample
 
 
 def extract_and_combine(output_dir, num_sims):
